fix(tests) resolve pb introduced by merge
